Model "monotone" payouts with --monotone
If the encrypting party is ok with giving away all secrets at a
higher (or lower) index than the oracle's attestation we can optimize
the verifiable encryption to only encrypt for the zero (or one) bits.

You enable this optimization with the --monotone cli flag and it only
does encryption padding for the 0 bits.